Add a new plugin(aaa_filter) to allow filtering on AAAA results
This new plugin allows you to resolve names as AAAA only if they are
within the cidr range of your network.

A use case is when you have to integrate an IPv6 network with an IPv4
only network with a hub and spoke model. In combination with the `dns64`
plugin this allows for local traffic to not go through any translation
but external traffic to use `dns64` exclusively.